+++ lib/rspec/core/option_parser.rb
@@ -63,12 +63,17 @@ module RSpec::Core
         end
 
         parser.on('--order TYPE[:SEED]', 'Run examples by the specified order type.',
-                  '  [default] files are ordered based on the underlying file',
-                  '            system\'s order',
+                  '  [defined] groups and examples are run in the order they are defined',
+                  '  [default] deprecated alias for defined',
                   '  [rand]    randomize the order of files, groups and examples',
                   '  [random]  alias for rand',
                   '  [random:SEED] e.g. --order random:123') do |o|
-          options[:order] = o
+          options[:order] = if o == 'default'
+                              RSpec.deprecate("RSpec's `--order default` CLI option", :replacement => "`--order defined`", :call_site => nil)
+                              'defined'
+                            else
+                              o
+                            end
         end
 
         parser.on('--seed SEED', Integer, 'Equivalent of --order rand:SEED.') do |seed|