+++ spec/rspec/core/configuration_spec.rb
@@ -1471,10 +1471,16 @@ module RSpec::Core
     end
 
     describe "#force" do
-      it "forces order" do
+      it "forces order (when given default)" do
         config.force :order => "default"
         config.order = "rand"
         expect(config.order).to eq("default")
+      end
+
+      it "forces order (when given defined)" do
+        config.force :order => "defined"
+        config.order = "rand"
+        expect(config.order).to eq("defined")
       end
 
       it "forces order and seed with :order => 'rand:37'" do
@@ -1486,7 +1492,7 @@ module RSpec::Core
 
       it "forces order and seed with :seed => '37'" do
         config.force :seed => "37"
-        config.order = "default"
+        config.order = "defined"
         expect(config.seed).to eq(37)
         expect(config.order).to eq("rand")
       end
@@ -1533,6 +1539,11 @@ module RSpec::Core
           expect(config.randomize?).to be_falsey
         end
       end
+
+      it 'is deprecated' do
+        expect_warn_deprecation_with_call_site(__FILE__, __LINE__ + 1, /randomize\?/)
+        config.randomize?
+      end
     end
 
     describe '#order=' do
@@ -1555,10 +1566,10 @@ module RSpec::Core
         end
       end
 
-      context 'given "default"' do
+      shared_examples_for "defined order" do |order|
         before do
           config.order = 'rand:123'
-          config.order = 'default'
+          config.order = order
         end
 
         it "sets the order to nil" do
@@ -1576,10 +1587,26 @@ module RSpec::Core
           expect(list.ordered).to eq([1, 2, 3, 4])
         end
       end
+
+      context "given 'default'" do
+        include_examples "defined order", 'default'
+
+        it 'prints a deprecation notice' do
+          expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /default/)
+          config.order = 'default'
+        end
+      end
+
+      context "given 'defined'" do
+        include_examples "defined order", 'defined'
+      end
     end
 
     describe "#order_examples" do
-      before { RSpec.stub(:configuration => config) }
+      before do
+        allow_deprecation
+        RSpec.stub(:configuration => config)
+      end
 
       it 'sets a block that determines the ordering of a collection extended with Extensions::Ordered::Examples' do
         examples = [1, 2, 3, 4]
@@ -1592,6 +1619,11 @@ module RSpec::Core
         config.order_examples { |examples| examples.reverse }
         expect(config.order).to eq("custom")
       end
+
+      it 'prints a deprecation warning' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /RSpec::Configuration#order_examples/)
+        config.order_examples { }
+      end
     end
 
     describe "#example_ordering_block" do
@@ -1601,7 +1633,10 @@ module RSpec::Core
     end
 
     describe "#order_groups" do
-      before { RSpec.stub(:configuration => config) }
+      before do
+        allow_deprecation
+        RSpec.stub(:configuration => config)
+      end
 
       it 'sets a block that determines the ordering of a collection extended with Extensions::Ordered::ExampleGroups' do
         groups = [1, 2, 3, 4]
@@ -1613,6 +1648,11 @@ module RSpec::Core
       it 'sets #order to "custom"' do
         config.order_groups { |groups| groups.reverse }
         expect(config.order).to eq("custom")
+      end
+
+      it 'prints a deprecation warning' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /RSpec::Configuration#order_groups/)
+        config.order_groups { }
       end
     end
 
@@ -1637,6 +1677,41 @@ module RSpec::Core
 
       it 'sets a block that determines the ordering of a collection extended with Extensions::Ordered::ExampleGroups' do
         expect(groups.ordered).to eq([4, 3, 2, 1])
+      end
+    end
+
+    describe "#register_ordering(:global)" do
+      let(:examples) { [1, 2, 3, 4].extend Extensions::Ordered::Examples }
+      let(:groups)   { [1, 2, 3, 4].extend Extensions::Ordered::ExampleGroups }
+
+      before do
+        RSpec.stub(:configuration => config)
+        config.register_ordering(:global) { |list| list.reverse }
+      end
+
+      it 'sets a block that determines the ordering of a collection extended with Extensions::Ordered::Examples' do
+        expect(examples.ordered).to eq([4, 3, 2, 1])
+      end
+
+      it 'sets a block that determines the ordering of a collection extended with Extensions::Ordered::ExampleGroups' do
+        expect(groups.ordered).to eq([4, 3, 2, 1])
+      end
+
+      it 'sets #order to "custom"' do
+        expect(config.order).to eq("custom")
+      end
+
+      it 'raises an ArgumentError if given a symbol other than `:global`' do
+        expect {
+          config.register_ordering(:custom) { }
+        }.to raise_error(ArgumentError, /:global/)
+      end
+    end
+
+    describe "#order" do
+      it 'is deprecated' do
+        expect_warn_deprecation_with_call_site(__FILE__, __LINE__ + 1, /order/)
+        config.order
       end
     end